requirements.txt
paramiko==3.4.0
mysql-connector-python==8.2.0
flask==3.0.0
pandas==2.1.4 